FIX SOME BUGS'

+++ src/main/java/com/anas/javautils/jls/output/FileInfo.java
@@ -71,7 +71,7 @@ public class FileInfo {
                 str = new ColoredString(String.format("%.2f %s", size, unit),
                         new TextColor.RGB(150, 153, 91));
             } else {
-                str = new ColoredString(String.format("%.1f %s", size, unit),
+                str = new ColoredString(String.format("%d %s", (int)size, unit),
                         new TextColor.RGB(150, 153, 91));
             }
         }
